pyproject.toml
[project]
name = "the-bus-factor"
version = "0.1.0"
description = "Bruin-powered weekly snapshot of open-source package continuity fragility (npm + PyPI)."
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "MIT" }
authors = [{ name = "The Bus Factor" }]

dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "duckdb>=1.1",
    "google-cloud-bigquery>=3.25",
    "tenacity>=8.3",
    "python-dateutil>=2.9",
    "structlog>=24.1",
    "pillow>=10.4",
    "jinja2>=3.1",
    "pyarrow>=16",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "sqlfluff>=3.1",
    "types-pyyaml>=6.0",
    "pandas-stubs>=2.2",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = ["pipeline/fixtures/*.duckdb"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
    "N",    # pep8-naming
]
ignore = [
    "E501",  # line too long — formatter handles it
]

[tool.ruff.lint.per-file-ignores]
# Bruin reads lowercase env vars (source_mode, duckdb_default); ignore SIM112's
# uppercase-constant rule for files that consume them.
"pipeline/assets/**/*.py" = ["SIM112"]
"pipeline/lib/live.py" = ["SIM112"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
files = ["pipeline", "tests"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
pythonpath = [".", "pipeline"]
